+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.10.2
 - Use compressed public keys by default
+- Generate Etherum wallets from Uint8List hashed entity address
   
 ## 0.10.1